feat: use MantineModal on various modals
